# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: UnderNet Safe Chat API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /upload:
    post:
      operationId: uploadFile
      tags: [upload]
      summary: Upload a file
      description: Upload an image or document file (multipart/form-data, field name "file")
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                  description: The file to upload (max 25MB)
      responses:
        "201":
          description: File uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttachmentResponse"
        "400":
          description: Invalid file
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /download/{id}:
    get:
      operationId: downloadFile
      tags: [upload]
      summary: Download a file
      description: Download an uploaded file by attachment ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: File content
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /sync/messages:
    get:
      operationId: syncMessages
      tags: [sync]
      summary: Sync missed messages
      description: Fetch messages since a given timestamp
      parameters:
        - name: since
          in: query
          schema:
            type: string
          description: Timestamp in milliseconds
      responses:
        "200":
          description: Synced messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyncMessagesResponse"
  /sync/conversations:
    get:
      operationId: syncConversations
      tags: [sync]
      summary: Sync conversation updates
      description: Fetch conversation updates since a given timestamp
      parameters:
        - name: since
          in: query
          schema:
            type: string
          description: Timestamp in milliseconds
      responses:
        "200":
          description: Synced conversations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyncConversationsResponse"
  /sync/status:
    get:
      operationId: syncStatus
      tags: [sync]
      summary: Sync read/delivery status
      description: Fetch status updates since a given timestamp
      parameters:
        - name: since
          in: query
          schema:
            type: string
          description: Timestamp in milliseconds
      responses:
        "200":
          description: Status updates
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyncStatusResponse"
  /notifications/subscribe:
    post:
      operationId: notificationSubscribe
      tags: [notifications]
      summary: Register push subscription
      description: Register a web push notification subscription
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotificationSubscribeBody"
      responses:
        "201":
          description: Subscription registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationSubscribeResponse"
        "200":
          description: Subscription already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationSubscribeResponse"
  /notifications/unsubscribe:
    delete:
      operationId: notificationUnsubscribe
      tags: [notifications]
      summary: Unsubscribe from push notifications
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotificationUnsubscribeBody"
      responses:
        "200":
          description: Unsubscribed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessStatusResponse"
...
  /storage/uploads/request-url:
    post:
      operationId: requestUploadUrl
      tags: [storage]
      summary: Request presigned upload URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestUploadUrlBody"
      responses:
        "200":
          description: Presigned URL generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequestUploadUrlResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    SuccessStatusResponse:
      type: object
      properties:
        status:
          type: string
      required:
        - status
...
    AttachmentInfo:
      type: object
      properties:
        id:
          type: integer
        originalName:
          type: string
        mimeType:
          type: string
        fileSize:
          type: integer
        storageKey:
          type: string
      required:
        - id
        - originalName
        - mimeType
        - fileSize
        - storageKey
    MessageResponse:
      type: object
      properties:
        id:
          type: integer
        clientMessageId:
          type: string
        conversationId:
          type: integer
        senderId:
          type: integer
        type:
          type: string
        textContent:
          type: ["string", "null"]
        attachmentId:
          type: ["integer", "null"]
        attachment:
          oneOf:
            - $ref: "#/components/schemas/AttachmentInfo"
            - type: "null"
        status:
          type: string
        deliveredAt:
          type: ["string", "null"]
          format: date-time
        readAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - clientMessageId
        - conversationId
        - senderId
        - type
        - status
        - createdAt
...
    AttachmentResponse:
      type: object
      properties:
        id:
          type: integer
        originalName:
          type: string
        mimeType:
          type: string
        fileSize:
          type: integer
        storageKey:
          type: string
      required:
        - id
        - originalName
        - mimeType
        - fileSize
        - storageKey
    SyncMessagesResponse:
      type: object
      properties:
        messages:
          type: array
          items:
            $ref: "#/components/schemas/MessageResponse"
      required:
        - messages
    SyncConversationItem:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
        participantA:
          type: integer
        participantB:
          type: integer
        lastMessageId:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - type
        - participantA
        - participantB
        - createdAt
        - updatedAt
    SyncConversationsResponse:
      type: object
      properties:
        conversations:
          type: array
          items:
            $ref: "#/components/schemas/SyncConversationItem"
      required:
        - conversations
    SyncStatusUpdate:
      type: object
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        status:
          type: string
        deliveredAt:
          type: ["string", "null"]
          format: date-time
        readAt:
          type: ["string", "null"]
          format: date-time
      required:
        - id
        - conversationId
        - status
    SyncStatusResponse:
      type: object
      properties:
        statusUpdates:
          type: array
          items:
            $ref: "#/components/schemas/SyncStatusUpdate"
      required:
        - statusUpdates
    NotificationSubscribeBody:
      type: object
      properties:
        subscription:
          type: object
      required:
        - subscription
    NotificationSubscribeResponse:
      type: object
      properties:
        id:
          type: integer
        status:
          type: string
      required:
        - id
        - status
    NotificationUnsubscribeBody:
      type: object
      properties:
        subscription:
          type: object
...
    RequestUploadUrlBody:
      type: object
      properties:
        name:
          type: string
        size:
          type: integer
        contentType:
          type: string
      required:
        - name
        - size
        - contentType
    RequestUploadUrlResponse:
      type: object
      properties:
        uploadURL:
          type: string
        objectPath:
          type: string
      required:
        - uploadURL
        - objectPath